requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=6.0
pytest-cov>=2.0
responses>=0.18.0

# Code quality
black>=21.0
flake8>=3.8
mypy>=0.800
isort>=5.0

# Documentation
mkdocs-material>=8.0
mkdocs-mermaid2-plugin>=0.6.0